Add van acceleration and game threads
The pedal now increases the acceleration of the van, depending on how much pressure is applied (how far up the depressed finger is). The houses, fences, and roads move in the opposite direction at the speed of the van, giving a movement effect. Game threads were added at a rate of 30 ticks per second, to allow for this.