Add the ability to throw parcels
Parcels can now be thrown from the van with a swipe beginning from the van, and ending anywhere on the screen. The parcel is animated, appearing to fly through the air and land behind the fence of a house. The parcel then travels left at the same rate as the houses and roads, until it moves off screen.