Add graphics to gap between streets
Shop and perpendicular street graphics have been added. They are rendered with one shop at the end of the first street, another shop drawn flipped at the start of the next street, with a perpendicular street going off into the distance to separate them. Shops have a space available for a random graphic to be generated, in the form of a shop banner related to the street name.