Add parcel contracts
At the beginning of each day, three randomly generated contracts of varying difficulty are shown to the player. These contracts consist of a number of parcels and target density, which is generated based on a targeted speed which increases each day, following a logarithmic curve. A score and penalty multiplier are still to be added. Once the player selects a contract, the day begins with the values from the chosen contract. The current target speed is only set for endless mode.