Add menu background animation
An animation has been added to the background of the main menu. A van enters from the left side of the screen, and slows down until it stops in the centre. A random number of parcels are then thrown from the left side of the screen into the van. Once all parcels are loaded, the van slowly speeds up and drives off the right side of the screen. The animation then repeats from the beginning.