Add functionality to main menu buttons
The buttons on the main menu can now all be interacted with, with the borders of a button turning white when it is touched. The two main buttons for Endless and Challenge mode now start the game, meaning the main game is now accessible once again. The exit button also performs its function, by exiting the application. The other buttons on the main menu do not yet link to their relevant menu screens.