Add game over menu
A game over menu has been added as part of the menus. This menu is accessed when the game finishes, either through loss of lives or end of days. Currently it displays the score from the game, and has a button for playing the same game mode again, along with another button to return to the main menu.