Adjust parcel throwing mechanics
Parcel throwing has been adjusted to only throw when the finger is moved at least a (small) certain distance, in order to avoid throwing multiple packages at once. The parcel landing region has also been adjusted, to ensure that it encompasses the same region to the left and the right of the house of about half a parcel's width.