Add parcel score text
When a parcel lands at the correct target, a green text pops up for a second, showing the amount of score that parcel has earned. The score text slowly moves upwards and fades as its time left on screen comes to a close. Additionally, fog has been added to the gap between shops, smoothing the transition between distant ground and sky.