Add level and experience gain
Experience is now earned at the end of the game, on the game over screen. Levels are also increased if the experience is greater than the next level's experience requirement. An animated level bar has been added on the game over screen to  show the experience increasing to the earned amount.