Add van unlocks and upgrades
Van styles are now locked behind levels, appearing shaded, and requiring a certain level to be selected. The previous method of upgrading vans was removed, due to the duality of acceleration and max speed reducing the number of real upgrade choices to two. The upgrades were instead made to be automatic with level, so that each new level makes the van slightly faster, and brakes slightly harder. Increased the detection area above the pedals.