Add sounds
Sounds have been recorded and implemented for both the game and the menu. In game, the van's engine now revs with a pitch that increases with speed. When a parcel lands at the correct target, a pleasing ding is played, whereas a miss results in a low note being played. In the menu, when a button a finger is lifted from a button, a click sound plays, coninciding with the execution of the button's task.